+++ website/__init__.py
@@ -10,11 +10,16 @@ def create_app():
     app = Flask(__name__)
     app.config['SECRET_KEY'] ='secretlol'
     app.config['SQLALCHEMY_DATABASE_URI'] = f"sqlite:///{DB_NAME}"
-
-    from models import User
-    create_db(app)
-
     db.init_app(app)
+
+    from .views import views
+    app.register_blueprint(views, url_prefix='/')
+
+
+    with app.app_context():
+        db.create_all()
+
+
 
     return app
